+++ CHANGELOG.md
@@ -10,6 +10,11 @@ All notable changes to this project will be documented in this file.
 - support for preprocessing CMIP data ([#401](https://github.com/WAM2layers/WAM2layers/pull/401)).
 - experimental support for parallel preprocessing ([#401](https://github.com/WAM2layers/WAM2layers/pull/401)).
 - config file documentation expanded ([#398](https://github.com/WAM2layers/WAM2layers/pull/398)).
+- support for preprocessing ARCO-ERA5 data ([#401](https://github.com/WAM2layers/WAM2layers/pull/401)).
+- longitude shifter to preprocessing that ensures all longitude values are in the range (-180, 180) ([#401](https://github.com/WAM2layers/WAM2layers/pull/401)).
+
+### Changed
+- input data to the tracking code is now assumed to have longitude values between -180 and 180 degrees ([#401](https://github.com/WAM2layers/WAM2layers/pull/401)).
 
 ### Fixed